requirements.txt
json5==0.9.14
numpy==1.21.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
plotly==5.15.0
#scipy==1.5.4
seaborn==0.11.2
Shapely==1.8.5.post1
#torch==1.10.2
tqdm==4.64.1
grpcio==1.32.0
grpcio_reflection==1.32.0
Pillow==9.0.0
protobuf==3.20.0